central/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
websockets==12.0
sqlmodel==0.0.19
openai==1.35.3
httpx==0.27.0
python-dotenv==1.0.1
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
aiosqlite==0.20.0
fpdf2==2.7.9
